More aggressively evaluate fixups targeting the .eh_frame section
On some targets (eg RISCV) fixups are not as aggressively
resolved and instead relocations are generated to be resolved
by the linker. This can cause a problem in the .eh_frame section,
as the parsing of the .eh_frame's CIEs and FDEs by ld uses the
raw values in the fields of the call frame information without
accounting for relocations.